+++ games_services/android/src/main/kotlin/com/abedalkareem/games_services/models/SavedGame.kt
@@ -3,5 +3,8 @@ package com.abedalkareem.games_services.models
 data class SavedGame(
   val name: String?,
   val modificationDate: Long?,
-  val deviceName: String?
+  val deviceName: String?,
+  val description: String? = null,
+  val playedTimeMillis: Long? = null,
+  val coverImage: String? = null
 )