+++ games_services_platform_interface/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 5.1.0
 
+- Add optional `coverImage`, `description`, and `playedTime` parameters to `saveGame`. (#228)
+- Add `showSavedGames` method to use Google Play Games default UI.
+- Include Google Play Games metadata with returned `SavedGame`s.
 - Add `timeScope` and `playerScope` parameters to `Leaderboards.showLeaderboards`.
 - Migrate to Swift Package Manager.
 - Migrate to built-in Kotlin (Bumps minimum supported SDK version to Flutter 3.44/Dart 3.12).